backend-py/requirements.txt
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0
selenium>=4.17.0
undetected-chromedriver>=3.5.5
fake-useragent>=1.5.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
openai>=1.12.0
anthropic>=0.18.0
